+++ CHANGELOG.md
@@ -2,4 +2,16 @@
 
 ## v0.1.1
 
-- [feat] support webpackPlugins、webpackLoaders、babelPlugins、babelPresets、postcssrc、postcssOptions to auto transform.
+- [feat] support some options to auto transform.
+  * webpackPlugins
+  * webpackLoaders
+  * babelPlugins
+  * babelPresets
+  * postcssrc
+  * postcssOptions
+  * cssLoaderOptions
+  * lessLoaderOptions
+  * sassLoaderOptions
+  * esbuild
+  * vendor
+  * modularImportRuntime